Revert "Feature/search list"
