+++ Testing/index.html
@@ -1,0 +1,63 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="styles.css">
+    <title>Star Wars Trading Card</title>
+</head>
+
+<body>
+    <div class="card" id="card1">
+        <div class="card-header">
+            <div class="circleID">
+                <span class="card-id" id="card-id1">1</span>
+            </div>
+            <span class="card-name" id="card-name1">Luke Skywalker</span>
+        </div>
+        <div class="card-img" id="card-img"></div>
+        <div class="card-content">
+            <div class="card-info">
+                <span class="info-label">Gender:</span>
+                <span class="info-value" id="gender-value1">male</span>
+            </div>
+            <div class="card-info">
+                <span class="info-label">Height:</span>
+                <span class="info-value" id="height-value1">172</span>
+            </div>
+            <div class="card-info">
+                <span class="info-label">Weight:</span>
+                <span class="info-value" id="weight-value1">70</span>
+            </div>
+            <div class="card-info">
+                <span class="info-label">Home Planet:</span>
+                <span class="info-value" id="home-planet-value1">MajMåne</span>
+            </div>
+            <div class="card-info">
+                <span class="info-label">Eye Color:</span>
+                <span class="info-value" id="eye-color-value1">blue</span>
+            </div>
+            <div class="card-info">
+                <span class="info-label">Hair Color:</span>
+                <span class="info-value" id="hair-color-value1">blonde</span>
+            </div>
+            <div class="card-info">
+                <span class="info-label">Skin Color:</span>
+                <span class="info-value" id="skin-color-value1">fair</span>
+            </div>
+            <div class="card-info">
+                <span class="info-label">Movie Amount:</span>
+                <span class="info-value" id="movie-amount-value1">4</span>
+            </div>
+            <!-- Add more info lines as needed -->
+        </div>
+        <div class="additional-info" id="additional-info">
+            <p>Person first appeared in move in this year and name and stuff.</p>
+            <p>Yes indeed more info</p>
+            <p> Their most expensive vehicle is "something" and costs "smth" credits</p>
+        </div>
+    </div>
+</body>
+
+</html>
