+++ Testing/styles.css
@@ -1,0 +1,77 @@
+body {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+    background-color: #ec8e8e;
+    margin: 0;
+    font-family: Arial, sans-serif;
+    background-color: rgb(103, 55, 55);
+}
+
+.card {
+    color: #f0f0f0;
+    display: flex;
+    flex-direction: column;
+    width: 300px;
+    background-color: rgba(42, 35, 35, 0.647);
+    border: 4px solid rgba(255, 255, 255, 0.414);
+    border-radius: 10px;
+    overflow: hidden;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+.card-header {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 1rem;
+    background-color: rgba(71, 62, 62, 0.8);
+}
+
+.card-id,
+.card-name {
+    font-weight: bold;
+}
+
+.card-img {
+    width: 100%;
+    height: 150px;
+    background-image: url("https://www.slashfilm.com/img/gallery/heres-what-the-new-luke-skywalker-looks-like-under-his-digital-mask/intro-1644193087.jpg");
+    background-position: center;
+    background-repeat: no-repeat;
+    background-size: cover;
+}
+
+.card-content {
+    padding: 1rem;
+}
+
+.card-info {
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 0.5rem;
+}
+
+.info-label {
+    font-weight: bold;
+}
+
+.additional-info {
+    padding: 1rem;
+    background-color: rgba(53, 45, 45, 0.8);
+    font-size: 0.9rem;
+    text-align: justify;
+}
+
+.circleID {
+    background-color: rgb(0, 8, 255);
+    border-radius: 16px;
+    padding: 5px;
+}
+
+.card-id {
+    padding-top: 10px;
+    padding: 5px;
+
+}
