+++ index.html
@@ -14,36 +14,19 @@
 </head>
 
 <body>
+    <header>
 
+
+        <!-- <img ng-src="./assets/img/characters/1.jpg" onerror="this.src='./assets/img/placeholder.jpg'"
+            src="./assets/img/characters/1.jpg">
+             -->
+    </header>
 
     <video id="loading-video" src="" autoplay loop muted></video>
 
 
     <main class="main">
-        <div class="header">
-
-            <div>
-                <div>
-                    <h1>Star</h1>
-                </div>
-                <div>
-                    <h1> Wars </h1>
-                </div>
-            </div>
-            <h2>Trivia</h2>
-            <div class="logos">
-                <a href="https://github.com/WilhelmErik/" target="_blank"><img
-                        src="./src/assets/images/logos/github-icon-logo.svg" alt="" height="28px"></a>
-                <a href="https://www.linkedin.com/in/albin-lundstr%C3%B6m/" target="_blank"><img
-                        src="./src/assets/images/logos/linkedin-icon-logo.svg" alt="" height="28px"></a>
-            </div>
-
-        </div>
-
-
-        <video id="background-video" autoplay muted loop>
-            <source src="./src/assets/videos/Star Wars - H264-compressed.mp4" type="video/mp4">
-        </video>
+        <h1>Star Wars Trivia</h1>
 
         <div class="people">
             <div id="person1" class="person">
