+++ src/scripts/main.js
@@ -1,7 +1,6 @@
 import { fetchCharacter } from "./modules/api-fetch.mjs";
-// import { Character } from "./modules/character.js";
-import { createCharacter } from "./modules/character.js";
-// import { compareNumeric } from "./character.js";
+import { createCharacter } from "./character.js";
+import { compareNumeric } from "./character.js";
 
 
 
@@ -23,7 +22,7 @@ clickButton.addEventListener("click", async function () {
         video.style.visibility = "visible";
         video.muted = false;
     }
-    //Kommentar
+
     let leftSelect = document.querySelector("select[name='select1']").value
     console.log(leftSelect, "👈")
     let rightSelect = document.querySelector("select[name='select2']").value
@@ -66,7 +65,6 @@ async function fetchAndPrint(obj1, obj2) {
         addInfo1.style.display = "none"
 
         let addInfo2 = document.getElementById("additional-info2")
-
         addInfo2.innerHTML = ""
         addInfo2.style.display = "none"
         //---
