+++ src/styles/sass/style.scss
@@ -1,66 +1,16 @@
 @font-face {
     font-family: 'StarWars';
-    src: url("../../assets/fonts/Starjhol.ttf") format('truetype');
+    src: url("../..//assets/fonts/Starjhol.ttf") format('truetype');
     font-weight: normal;
     font-style: normal;
 }
 
-@font-face {
-    font-family: 'Grotesk';
-    src: url("../../assets/fonts/Typo Grotesk Demo.otf") format('opentype');
-    font-weight: normal;
-    font-style: normal;
-}
+$cardColor: rgba(6, 2, 2, 0.504);
 
-
-
-$cardColor: rgba(3, 0, 0, 0.504);
-$SWYellow : #ffe919;
-
-.header {
-    background-color: rgba(0, 0, 0, 0.026);
-    height: 15vh;
-    margin: 0;
-    padding: 0;
-    display: flex;
-    justify-content: center;
-
-}
-
-
-
-h1 {
-    margin-top: -20px;
-    margin-bottom: -35px;
-    text-align: center;
-    font-family: 'StarWars', sans-serif;
-    font-weight: 900;
-    font-size: 60px;
-    color: $SWYellow;
-}
-
-h2 {
-    font-family: 'StarWars', sans-serif;
-    margin-top: 20px;
-    color: $SWYellow;
-    margin-left: 15px;
-    font-size: 48px;
-}
-
-#background-video {
-    object-fit: cover;
-    width: 100%;
-    height: 100%;
-    z-index: -1;
-    position: absolute;
-    left: 0;
-    top: 0;
-}
 
 
 
 .people {
-    min-height: 75vh;
 
     .person {
         display: flex;
@@ -108,70 +58,50 @@ h2 {
     box-shadow: 1px 5px;
     font-family: 'StarWars';
     background-color: rgba(0, 0, 0, 0.052);
-    color: $SWYellow;
+    color: yellow;
 }
 
 main {
-    // display: flex;
-    // flex-direction: column;
-    // justify-content: center;
-
-    // ----------
-    margin-top: 0;
     text-align: center;
-    position: relative;
-    // overflow: hidden;
-    background: linear-gradient(to top, #00000081, #030303);
-    // height: 85vh;
 
+    background: linear-gradient(to top, #1b1329ba, #08060a82), url("../../assets/images/background1.gif");
+    background-repeat: no-repeat;
+    background-size: cover;
+    background-position: center 30%;
 }
 
 
-
+h1 {
+    font-family: 'StarWars', sans-serif;
+    font-weight: 900;
+    font-size: 42px;
+    color: rgb(186, 178, 30);
+}
 
 select {
-    max-height: 50px;
-    border-radius: 12px;
-    font-size: 24px;
+    font-size: 18px;
     padding: 5px 10px;
     // font-family: StarWars;
     color: aliceblue;
     background-color: black;
-    color: $SWYellow;
-    overflow-y: scroll;
 }
-
-.select-test {
-    margin: auto;
-}
-
-
 
 body {
-    margin: 0px;
-    padding: 0px;
-
+    background-color: rgb(6, 5, 5);
+    height: 100vh;
 }
 
 #loading-check {
-    // border-radius: 40px;
     display: flex;
     font-size: 24px;
     color: aliceblue;
-    background-color: $cardColor;
+    background-color: rgb(12, 10, 10);
     justify-content: center;
     align-items: center;
 
-    margin: auto;
-    // margin-top: 5vh;
-
-
-
-
     input.loading-check {
-        height: 30px;
-        width: 30px;
-
+        height: 20px;
+        width: 20px;
 
     }
 }
@@ -192,15 +122,11 @@ body {
 }
 
 #stats {
-    font-family: 'Grotesk', sans-serif;
     border-radius: 50px;
     align-self: center;
     min-height: 50%;
-    padding: 1rem;
     backdrop-filter: blur(6px);
-    visibility: hidden;
-    background-color: $cardColor;
-    border: 4px groove rgba(255, 255, 255, 0.475);
+    // background-color: rgba(12, 10, 7, 0);
 
     p {
         font-size: 20px;
@@ -208,16 +134,6 @@ body {
     }
 }
 
-#more-text-btn {
-    margin: 1rem;
-    padding: 1rem;
-    font-size: 16px;
-    border-radius: 2rem;
-    border-style: inset;
-    box-shadow: 1px 5px;
-    background-color: rgba(0, 0, 0, 0.668);
-    color: yellow;
-}
 
 .profile-picture {
     margin: auto;
@@ -264,15 +180,8 @@ body {
         grid-template-rows: 1fr;
         grid-template-columns: 1fr; // background-color: rgb(10, 10, 10);
         color: aliceblue;
-
     }
 
-    main {
-        overflow-y: auto;
-        background: linear-gradient(to top, #1b1329ba, #08060a82), url("../../assets/images/background1.gif");
-        background-repeat: no-repeat;
-        background-size: cover;
-    }
 }
 
 
@@ -280,7 +189,9 @@ body {
 
 
 @media screen and (max-width: 600px) {
-    body {}
+    body {
+        background-color: olive;
+    }
 
     #loading-video {
         position: fixed;
@@ -298,7 +209,6 @@ body {
         // display: none;
     }
 
-    main {}
 }
 
 
@@ -329,14 +239,14 @@ body {
     max-width: 358px;
     width: 100%;
     background-color: $cardColor;
-    border: 4px groove rgba(255, 255, 255, 0.475);
+    border: 4px inset rgba(255, 255, 255, 0.475);
     border-radius: 10px;
     overflow: hidden;
     box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
     margin: auto;
     display: none;
-    backdrop-filter: blur(6px);
-
+     backdrop-filter: blur(6px);
+    margin-top: 10%;
 }
 
 .card-header {
@@ -375,7 +285,6 @@ body {
 }
 
 .card-info {
-    font-family: 'Grotesk', sans-serif;
     font-size: 20px;
     display: flex;
     justify-content: space-between;
@@ -394,7 +303,7 @@ body {
     text-align: justify;
     display: none;
     border-top: gray inset 2px;
-
+   
 }
 
 .circleID {
