+++ src/styles/style.css
@@ -1,0 +1,17 @@
+.people {
+
+    display: grid;
+    grid-template-rows: minmax(0, auto);
+    grid-template-columns: minmax(auto, 50%) 1fr;
+    background-color: rgba(0, 26, 255, 0.605);
+    /* grid-column: 1fr, 1fr; */
+}
+
+.people div {
+    border: rgb(25, 11, 11) 2px dotted;
+
+
+
+}
+
+/*# sourceMappingURL=style.css.map */
